+++ scripts/release/install.js
@@ -10,9 +10,9 @@ const out = "./dist/installers"
 const appPath = "dist/packages/Brim-darwin-x64/Brim.app"
 
 module.exports = {
-  darwin: function() {
+  darwin: async function() {
     console.log("Building installer for darwin")
-    createDMG(
+    await createDMG(
       {
         overwrite: true,
         appPath,
@@ -20,21 +20,21 @@ module.exports = {
         out
       },
       (err) => {
-        if (!err) {
-          console.log("Built installer for darwin in " + out)
-        } else {
-          console.log("Error builing darwin installer " + err)
+        if (err) {
+          throw new Error("Error builing darwin installer " + err)
         }
+        console.log("Built installer for darwin in " + out)
       }
     )
-    createZip(
+
+    await createZip(
       {
         dir: appPath,
         out: path.join(out, "Brim-darwin-autoupdater.zip")
       },
       (err) => {
         if (err) {
-          console.log("Error zipping darwin package: " + err)
+          throw new Error("Error zipping darwin package: " + err)
         }
         console.log("Zip for darwin package written in " + out)
       }
